# docs/apidog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FundLedger API",
    "version": "0.1.0",
    "description": "Apidog-importable OpenAPI collection for the FundLedger multi-tenant SaaS API."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.fundledger.app",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/members": {
      "get": {
        "tags": ["Members"],
        "summary": "List members",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdHeader" },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/MemberStatus" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Members"],
        "summary": "Add member and validate share cap",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/MemberCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/v1/members/{id}": {
      "get": {
        "tags": ["Members"],
        "summary": "Member detail with contribution summary",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/Error" } }
      },
      "patch": {
        "tags": ["Members"],
        "summary": "Update member contact, shares or status",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "requestBody": { "$ref": "#/components/requestBodies/MemberUpdate" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/schedules": {
      "get": {
        "tags": ["Schedules"],
        "summary": "List schedules",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Schedules"],
        "summary": "Create schedule and generate fixed per-member dues",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ScheduleCreate" },
        "responses": { "201": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/recurring-schedules": {
      "get": {
        "tags": ["Schedules"],
        "summary": "List recurring schedule rules",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "post": {
        "tags": ["Schedules"],
        "summary": "Create recurring schedule rule",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecurringScheduleCreate" },
              "example": {
                "name": "Service Charge",
                "unit_amount": 3000,
                "frequency": "monthly",
                "starts_on": "2026-08-01"
              }
            }
          }
        },
        "responses": { "201": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/recurring-schedules/{id}": {
      "patch": {
        "tags": ["Schedules"],
        "summary": "Update recurring schedule rule",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecurringScheduleUpdate" },
              "example": { "unit_amount": 3500, "frequency": "quarterly" }
            }
          }
        },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "tags": ["Schedules"],
        "summary": "Deactivate recurring schedule rule",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/v1/schedules/{id}": {
      "patch": {
        "tags": ["Schedules"],
        "summary": "Edit or close schedule",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectIdHeader" }, { "$ref": "#/components/parameters/IdPath" }],
        "requestBody": { "$ref": "#/components/requestBodies/ScheduleUpdate" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/v1/penalty-policy": {
      "get": {
        "tags": ["Penalties"],
        "summary": "Resolve effective penalty policy",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdHeader" },
          { "name": "schedule_id", "in": "query", "schema": { "type": "string" } },
          { "name": "scope", "in": "query", "schema": { "type": "string", "enum": ["client", "project"] } }
        ],
        "responses": { "200": { "$ref": "#/components/responses/Success" } }
      },
      "put": {
        "tags": ["Penalties"],
        "summary": "Set client or project penalty policy",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/ProjectIdHeader" },
          { "name": "scope", "in": "query", "schema": { "type": "string", "enum": ["client", "project"], "default": "project" } }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/PenaltyPolicy" },
        "responses": { "200": { "$ref": "#/components/responses/Success" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ProvisioningApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Billing/provisioning secret configured via PROVISIONING_API_KEYS."
      },
      "ProvisioningAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Alternative form: Authorization: ApiKey <key>."
      }
    },
    "parameters": {
      "ProjectIdHeader": {
        "name": "X-Project-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Active project scope. If omitted, the session active project is used."
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      }
    },
    "responses": {
      "Success": {
        "description": "Success response envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/SuccessEnvelope" }
          }
        }
      },
      "Error": {
        "description": "Error response envelope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" }
          }
        }
      }
    },
    "requestBodies": {
...
      "MemberCreate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/MemberCreate" },
            "example": { "name": "Imran Hossain", "mobile": "+8801911553300", "shares": 2, "address": "Flat B-4" }
          }
        }
      },
      "MemberUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/MemberUpdate" },
            "example": { "shares": 3, "status": "active" }
          }
        }
      },
      "ScheduleCreate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ScheduleCreate" },
            "example": { "name": "Roofing Assessment", "unit_amount": 50000, "due_date": "2026-11-20" }
          }
        }
      },
      "ScheduleUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ScheduleUpdate" },
            "example": { "status": "closed" }
          }
        }
      },
...
      "PenaltyPolicy": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/PenaltyPolicy" },
            "example": { "enabled": true, "grace_months": 1, "onetime_percent": 10 }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "ok": { "type": "boolean", "example": true },
          "data": { "type": "object", "additionalProperties": true },
          "meta": { "type": "object", "additionalProperties": true }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "VALIDATION" },
              "message": { "type": "string", "example": "Invalid request" },
              "fields": { "type": "object", "additionalProperties": true }
            }
          }
        }
      },
...
      "MemberStatus": {
        "type": "string",
        "enum": ["active", "inactive"]
      },
      "ScheduleStatus": {
        "type": "string",
        "enum": ["draft", "active", "closed"]
      },
...
      "RecurringFrequency": {
        "type": "string",
        "enum": ["weekly", "biweekly", "monthly", "bimonthly", "quarterly", "yearly"]
      },
...
      "PenaltyPolicy": {
        "type": "object",
        "required": ["enabled", "grace_months"],
        "properties": {
          "enabled": { "type": "boolean" },
          "grace_months": { "type": "integer", "minimum": 0, "maximum": 2 },
          "onetime_percent": { "type": "number", "minimum": 0, "maximum": 100 },
          "recurring_percent": { "type": "number", "minimum": 0, "maximum": 100 },
          "recurring_period": { "type": "string", "enum": ["weekly", "monthly"] },
          "compounding": { "type": "boolean" },
          "max_percent": { "type": "number", "minimum": 0, "maximum": 100 }
        }
      },
...
      "MemberCreate": {
        "type": "object",
        "required": ["name", "mobile", "shares"],
        "properties": {
          "name": { "type": "string", "minLength": 2 },
          "mobile": { "type": "string", "minLength": 6 },
          "shares": { "type": "integer", "minimum": 1 },
          "address": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "previous_due_amount": { "type": "integer", "minimum": 0 }
        }
      },
      "MemberUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "mobile": { "type": "string" },
          "shares": { "type": "integer", "minimum": 1 },
          "address": { "type": "string", "nullable": true },
          "email": { "type": "string", "format": "email", "nullable": true },
          "status": { "$ref": "#/components/schemas/MemberStatus" }
        }
      },
      "ScheduleCreate": {
        "type": "object",
        "required": ["name", "unit_amount", "due_date"],
        "properties": {
          "name": { "type": "string" },
          "unit_amount": { "type": "integer", "minimum": 1 },
          "due_date": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/ScheduleStatus" },
          "penalty_policy": { "$ref": "#/components/schemas/PenaltyPolicy" }
        }
      },
      "RecurringScheduleCreate": {
        "type": "object",
        "required": ["name", "unit_amount", "frequency", "starts_on"],
        "properties": {
          "name": { "type": "string", "minLength": 2 },
          "unit_amount": { "type": "integer", "minimum": 1 },
          "frequency": { "$ref": "#/components/schemas/RecurringFrequency" },
          "starts_on": { "type": "string", "format": "date" },
          "is_active": { "type": "boolean", "default": true },
          "penalty_policy": { "$ref": "#/components/schemas/PenaltyPolicy" }
        }
      },
      "RecurringScheduleUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2 },
          "unit_amount": { "type": "integer", "minimum": 1 },
          "frequency": { "$ref": "#/components/schemas/RecurringFrequency" },
          "starts_on": { "type": "string", "format": "date" },
          "is_active": { "type": "boolean" },
          "penalty_policy": { "$ref": "#/components/schemas/PenaltyPolicy" }
        }
      },
      "ScheduleUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "due_date": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/ScheduleStatus" },
          "penalty_policy": { "$ref": "#/components/schemas/PenaltyPolicy" }
        }
      },
...
    }
  }
}